+++ R/FindOptimalSubset.R
@@ -38,9 +38,8 @@
 #'   Number of consecutive generations without any improvement in the
 #'   \dQuote{best} fitness value before the GA is stopped.
 #' @param suggestions 'matrix'.
-#'   Integer (or binary) representation of chromosomes to be included in the initial population (optional).
-#'   For binary representation of chromosomes, the number of columns must match the number of decision variables.
-#'   See returned list components \code{solution} and \code{ga_output@solution} for suggested values for this arugment.
+#'   Integer chromosomes to be included in the initial population.
+#'   See returned \code{solution} component for a suggested value for this arugment.
 #' @param parallel 'logical' or 'integer'.
 #'   Whether to use parallel computing.
 #'   This argument can also be used to specify the number of cores
@@ -104,12 +103,11 @@
 #' # genearated from a standard uniform distribution.
 #' k <- 4
 #' n <- 100
-#' seed <- 321
+#' seed <- 123
 #' set.seed(seed); numbers <- sort(runif(n))
 #' Fitness <- function(string, n, numbers) {
 #'   idxs <- DecodeChromosome(string, n)
-#'   value <- -sum(numbers[idxs])
-#'   return(value)
+#'   -sum(numbers[idxs])
 #' }
 #' \dontrun{
 #' out <- FindOptimalSubset(n, k, Fitness, numbers, elitism = 1, run = 10,
@@ -139,7 +137,7 @@ FindOptimalSubset <- function(n, k, Fitness, ..., popSize=100,
   checkmate::assertInt(elitism, lower=0, upper=popSize)
   checkmate::assertInt(maxiter, lower=1)
   checkmate::assertInt(run, lower=1, upper=maxiter)
-  checkmate::assertMatrix(suggestions, null.ok=TRUE)
+  checkmate::assertMatrix(suggestions, min.rows=1, min.cols=1, null.ok=TRUE)
   checkmate::qassert(parallel, c("B1", "X1[0,)"))
   checkmate::assertFunction(monitor, null.ok=TRUE)
   if (is.null(monitor)) monitor <- FALSE
@@ -152,26 +150,20 @@ FindOptimalSubset <- function(n, k, Fitness, ..., popSize=100,
     numIslands <- parallel
 
   # calculate number of bits in the binary string representing the chromosome
-  nBits <- ceiling(log2(n + 1)) * k
+  nBits <- .CountBits(n) * k
 
   # format suggested chromosomes
   if (!is.null(suggestions)) {
-    if (identical(as.vector(suggestions), as.numeric(as.logical(suggestions)))) {
-      if (ncol(suggestions) != nBits)
-        stop("Problem with number of columns in binary 'suggestions' argument")
-    } else {
-      m <- suggestions
-      if (k < ncol(m)) {
-        set.seed(seed); m <- t(apply(m, 1, sample, size=k))
-      } else if (k > ncol(m)) {
-        idxs <- seq_len(n)
-        set.seed(seed)
-        m <- t(apply(m, 1, function(i) {
-          c(i, sample(idxs[-i], k - ncol(m)))
-        }))
-      }
-      suggestions <- t(apply(m, 1, function(i) EncodeChromosome(i, n)))
+    m <- suggestions
+    if (any(m > n)) stop("'suggestions' element value(s) greater than 'k'")
+    if (k < ncol(m)) {
+      m <- m[, seq_len(k), drop=FALSE]
+    } else if (k > ncol(m)) {
+      set.seed(seed)
+      m <- cbind(m, t(apply(m, 1, function(x) sample(seq_len(n)[-x], k - ncol(m)))))
     }
+    suggestions <- t(apply(m, 1, function(x) EncodeChromosome(x, n)))
+    stopifnot(nBits == ncol(suggestions))
   }
 
   # solve genetic algorithm
@@ -212,18 +204,14 @@ FindOptimalSubset <- function(n, k, Fitness, ..., popSize=100,
   solution <- m[!duplicated(m), , drop=FALSE]
 
   # bundle output
-  return(list(call=match.call(),
-              solution=solution,
-              ga_output=ga_output,
-              ga_time=ga_time))
+  list(call=match.call(), solution=solution, ga_output=ga_output, ga_time=ga_time)
 }
 
 .Population <- function(object, n) {
-  k <- object@nBits / ceiling(log2(n + 1))
+  k <- object@nBits / .CountBits(n)
   BuildChromosomes <- function(x) sample.int(n, k)
   m <- do.call("rbind", lapply(seq_len(object@popSize), BuildChromosomes))
-  pop <- t(apply(m, 1, function(i) EncodeChromosome(i, n)))
-  return(pop)
+  t(apply(m, 1, function(i) EncodeChromosome(i, n)))
 }
 
 .Mutate <- function(object, parent, n) {
@@ -240,8 +228,7 @@ FindOptimalSubset <- function(n, k, Fitness, ..., popSize=100,
     x_sorted <- sort(x)
     if (!any(apply(m, 1, function(y) identical(y, x_sorted)))) break
   }
-  mut <- EncodeChromosome(x, n)
-  return(mut)
+  EncodeChromosome(x, n)
 }
 
 .Crossover <- function(object, parents, n) {
@@ -265,10 +252,10 @@ FindOptimalSubset <- function(n, k, Fitness, ..., popSize=100,
   }))
   m <- t(apply(object@population, 1, function(i) sort(DecodeChromosome(i, n))))
   FindFitness <- function(child) {
-    return(object@fitness[which(apply(m, 1, function(i) identical(i, child)))[1]])
+    object@fitness[which(apply(m, 1, function(i) identical(i, child)))[1]]
   }
   fitness_children <- c(FindFitness(sort(c1)), FindFitness(sort(c2)))
-  return(list(children=encoded_children, fitness=fitness_children))
+  list(children=encoded_children, fitness=fitness_children)
 }
 
 
@@ -308,18 +295,25 @@ FindOptimalSubset <- function(n, k, Fitness, ..., popSize=100,
 #'
 
 EncodeChromosome <- function(x, n) {
-  width <- ceiling(log2(n + 1))
-  return(unlist(lapply(x, function(i) {
+  width <- .CountBits(n)
+  unlist(lapply(x, function(i) {
     GA::decimal2binary(i, width)
-  })))
+  }))
 }
 
 #' @rdname EncodeChromosome
 #' @export
 
 DecodeChromosome <- function(y, n) {
-  width <- ceiling(log2(n + 1))
-  return(vapply(seq(1, length(y), by=width), function(i) {
+  width <- .CountBits(n)
+  vapply(seq(1, length(y), by=width), function(i) {
     GA::binary2decimal(y[i:(i + width - 1L)])
-  }, 0))
+  }, 0)
+}
+
+
+# Count number of bits in a number
+
+.CountBits <- function(n) {
+  as.integer(floor(log2(n)) + 1)
 }